% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltypes.R
\name{cell_signatures}
\alias{cell_signatures}
\title{Cell type signatures}
\format{
An object of class tmodGS
}
\source{
CIBERSORT, CellMarkers, PanglaoDB
}
\description{
Cell type signatures
}
\examples{
## to use cell signatures, type
data(cell_signatures)
data(vaccination)
gl <- vaccination$GeneName[ order(vaccination$qval.F.D1) ]
tmodCERNOtest(gl, mset=cell_signatures)
}
