% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,tmod-method}
\alias{[,tmod,ANY,ANY,ANY-method}
\alias{[,tmod-method}
\alias{show,tmodSummary-method}
\title{Shows the tmod object}
\usage{
\S4method{show}{tmod}(object)

\S4method{[}{tmod,ANY,ANY,ANY}(x, i)

\S4method{show}{tmodSummary}(object)
}
\arguments{
\item{object}{a tmod object}

\item{x}{a tmod object}

\item{i}{indices specifying elements to extract or replace}
}
\description{
Shows the tmod object

Extracts parts of a tmod object

Shows a tmodSummary object
}
