% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{gg_evidencePlot}
\alias{gg_evidencePlot}
\title{Create an evidence plot for a module (ggplot2 version)}
\usage{
gg_evidencePlot(
  l,
  m,
  mset = NULL,
  filter = FALSE,
  unique = TRUE,
  gene.labels = NULL,
  gene.colors = NULL
)
}
\arguments{
\item{l}{sorted list of HGNC gene identifiers}

\item{m}{character vector of modules for which the plot should be created}

\item{mset}{Which module set to use (see tmodUtest for details)}

\item{filter}{if TRUE, genes not defined in the module set will be removed}

\item{unique}{if TRUE, duplicates will be removed}

\item{gene.labels}{if TRUE, gene names are shown; alternatively, a named character vector with gene labels to be shown, or NULL (default) for no labels (option evaluated only if rug is plotted)}

\item{gene.colors}{NULL (default) or a character vectors indicating the color for each gene. Either a named vector or a vector with the same order of genes as `l`.}
}
\description{
Create an evidence plot for a module (ggplot2 version)
}
