% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mbind}
\alias{mbind}
\title{Bind two or more transcriptional module sets}
\usage{
mbind(x, ...)
}
\arguments{
\item{x}{First module to bind}

\item{...}{further modules will be concatenated with x}
}
\value{
an object of class tmod which is the result of binding together
        all modules
}
\description{
Bind two or more transcriptional module sets
}
