% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modstats.R
\name{modcors}
\alias{modcors}
\title{Module correlation}
\usage{
modcors(x, g, mset = NULL, ...)
}
\arguments{
\item{x}{a data set, with variables (e.g. genes) in rows and samples in columns}

\item{g}{a vector of variable idenitifiers which correspond to the definition of modules}

\item{mset}{a module set}

\item{...}{any further parameters will be passed to the cor() function}
}
\value{
a matrix of module correlation coefficients
}
\description{
Calculate the correlation between modules
}
\details{
The correlation between modules are defined as 
correlation coefficient between the modules eigengenes.
These are based on a particular gene expression data set.

This function is a simple wrapper combining eigengene() and cor().
}
