% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modstats.R
\name{modjaccard}
\alias{modjaccard}
\title{Jaccard index for modules}
\usage{
modjaccard(mset = NULL, g = NULL)
}
\arguments{
\item{mset}{set of modules for which to calculate the Jaccard index. If
NULL, the default tmod module set will be used.}

\item{g}{a list of genes. If NULL, the list of genes from the mset will be
used.}
}
\value{
matrix with Jaccard index for each pair of modules in mset
}
\description{
Jaccard index for modules
}
\details{
For each pair of modules in mset, calculate the Jacard index between
these modules.
}
