% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{pcaplot}
\alias{pcaplot}
\title{Plot a PCA object returned by prcomp}
\usage{
pcaplot(pca, components = 1:2, group = NULL, col = "black",
  pch = 19, ...)
}
\arguments{
\item{pca}{PCA object returned by prcomp}

\item{components}{a vector of length two indicating the components to plot}

\item{group}{a factor determining shapes of the points to show (unless
overriden by pch=...)}

\item{col}{Color for plotting (default: grey)}

\item{pch}{Type of character to plot (default: 19)}

\item{...}{any further parameters will be passed to the plot() function
(e.g. col, cex, ...)}
}
\value{
If group is NULL, then NULL; else a data frame containing
        colors and shapes matching each group
}
\description{
Plot a PCA object returned by prcomp
}
