% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{renameMods}
\alias{renameMods}
\title{Rename module IDs}
\usage{
renameMods(mset, newids)
}
\arguments{
\item{mset}{a module set (object of class tmod)}

\item{newids}{a named vector of unique new IDs}
}
\value{
object of class tmod with renamed IDs
}
\description{
Change the IDs of modules
}
\details{
Rename the modules by replacing the selected IDs with new IDs.
}
