% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{showGene}
\alias{showGene}
\title{A combined beeswarm / boxplot}
\usage{
showGene(data, group, main = "", pch = 19, xlab = "",
  ylab = "log2 expression", las = 2, pwcol = NULL, ...)
}
\arguments{
\item{data}{a vector of numeric values to be plotted}

\item{group}{factor describing the groups}

\item{main}{title of the plot}

\item{pch}{character to plot the points}

\item{xlab, ylab}{x and y axis labels}

\item{las}{see par()}

\item{pwcol}{colors of the points (see beeswarm)}

\item{...}{any additional parameters to be passed to the beeswarm command}
}
\description{
A combined beeswarm / boxplot
}
\details{
This is just a simple wrapper around the beeswarm() and boxplot()
commands.
}
\examples{
data(Egambia)
E <- as.matrix(Egambia[,-c(1:3)])
showGene(E["20799",], rep(c("CTRL", "TB"), each=15))
}
