% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showGene.R
\name{showModule}
\alias{showModule}
\title{Select genes belonging to a module from a data frame}
\usage{
showModule(df, genes, module, mset = "all")
}
\arguments{
\item{df}{a data frame}

\item{genes}{a character vector with gene IDs}

\item{module}{a single character value, ID of the module to be shown}

\item{mset}{Module set to use; see "tmodUtest" for details}
}
\description{
Select genes belonging to a module from a data frame
}
\details{
showModule filters a data frame such that only genes from a module are
shown.
}

