% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panelplots.R
\name{simplePie}
\alias{simplePie}
\alias{simpleRug}
\title{Simple Pie Chart}
\usage{
simplePie(x, y, w, h, v, col, res = 100)

simpleRug(x, y, w, h, v, col)
}
\arguments{
\item{x,y}{coordinates at which to draw the plot}

\item{w,h}{width and height of the plot}

\item{v}{sizes of the slices}

\item{col}{colors of the slices}

\item{res}{resolution (number of polygon edges in a full circle)}
}
\description{
The simplePie function draws a simple pie chart at specified coordinates with
specified width, height and color. The simpleRug function draws a
corresponding rug plot.
}
\details{
simplePie() draws a pie chart with width w and height h at coordinates
(x,y). The size of the slices is taken from the numeric vector v, and
their color from the character vector col.
}
\examples{
plot(NULL, xlim=1:2, ylim=1:2)
col <- c("#cc000099", "#0000cc99")
for(i in 1:25) {
  x <- runif(1) + 1
  y <- runif(1) + 1
  simplePie( x, y, 0.05, 0.05, c(x,y), col)
}
}

