% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R, R/tmod-data.R
\docType{class}
\name{tmod-class}
\alias{tmod}
\alias{tmod-class}
\alias{tmod-data}
\title{An S4 class that stores modules.}
\description{
An S4 class that stores modules.

Gene expression module data
}
\details{
This data set contains the module annotations (tmod$MODULES), gene to
module (tmod$GENES2MODULES) and module to gene (tmod$MODULES2GENES) mappings
and a gene list (tmod$GENES).

tmod$MODULES and tmod$GENES are data frames, while tmod$MODULES2GENES and
tmod$GENES2MODULES are lists with, respectively, module and gene
identifiers as names.
}
\examples{
# list of first 10 modules
data(tmod)
tmod
tmod$MODULES[1:10, ]
tmod[1:10]
}

