% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmod-package.R
\name{tmod-package}
\alias{tmod-package}
\title{Transcriptional Module Analysis}
\description{
Transcriptional Module Analysis
}
\details{
The primary role of this package is to provide published module
assignments between genes and transcriptional modules, as well as tools
to analyse and visualize the modules.
}
\seealso{
\code{\link{tmodHGtest}}, \code{\link{tmodUtest}}
}
