% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes2.R
\name{tmod2tmodGS}
\alias{tmod2tmodGS}
\title{Convert the old tmod objects to the tmodGS objects}
\usage{
tmod2tmodGS(x)
}
\arguments{
\item{x}{an object of class tmod}
}
\value{
Returns an object of class tmodGS.
}
\description{
Convert the old tmod objects to the tmodGS objects
}
\details{
The old tmod representation was very inefficient. This function converts
the objects to a new representation which is both allowing faster
computations and more memory efficient.
}
