% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes2.R
\name{tmod_ids}
\alias{tmod_ids}
\alias{tmod_ids<-}
\alias{tmod_titles}
\alias{tmod_titles<-}
\title{Query and set IDs of gene sets in a tmodGS object}
\usage{
tmod_ids(x)

tmod_ids(x) <- value

tmod_titles(x)

tmod_titles(x) <- value
}
\arguments{
\item{x}{an object of class tmodGS}

\item{value}{a character vector of unique IDs}
}
\value{
Returns character vector corresponding to x$gs$ID
}
\description{
Query and set IDs (tmod_id) or Titles (tmod_title) of gene sets in a tmodGS object
}
\examples{
data(tmod)
mset <- tmod[ c("LI.M37.0", "LI.M75", "LI.M3") ]
tmod_ids(mset)
tmod_ids(mset) <- c("em", "pstrem", "bzdrem")
tmod_titles(mset) <- c("foo", "bar", "baz")
mset$gs
}
