% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lem.R
\name{tmodLEASummary}
\alias{tmodLEASummary}
\title{Summary stats of a leading edge analysis}
\usage{
tmodLEASummary(lea, genes = FALSE, labels = NULL, mset = NULL)
}
\arguments{
\item{lea}{result of `tmodLEA`}

\item{genes}{if TRUE, the gene identifiers of the leading edge (joined by commas) will be appended.}

\item{labels}{labels to add to the result; if NULL, the labels will be
taken from `mset`}

\item{mset}{Which module set to use. Either a character vector ("LI", "DC" or "all", default: all) or an object of class tmod (see "Custom module definitions" below)}
}
\value{
data frame with summary stats
}
\description{
Summary stats of a leading edge analysis
}
