% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{tmodPal}
\alias{tmodPal}
\title{A selection of color palettes}
\usage{
tmodPal(n = NULL, set = "friendly", alpha = 0.7, func = FALSE)
}
\arguments{
\item{n}{Number of colors to return (default: all for "friendly", 3 for everything else)}

\item{set}{Which palette set (see Details).}

\item{alpha}{0 for maximum transparency, 1 for no transparency.}

\item{func}{if TRUE, the returned object will be a function rather than
a character vector}
}
\value{
Either a character vector, or, when the func parameter is TRUE,
        a function that takes only one argument (a single number)
}
\description{
Return a preset selection of colors, adjusted by alpha
}
\details{
A few palettes have been predefined in tmod, and this function can be
used to extract them. The following palettes have been defined:
* friendly -- a set of distinct, colorblind-friendly colors
* bwr, rwb, ckp, pkc -- gradients (b-blue, r-red, w-white, c-cyan, k-blacK, p-purple)
By default, either all colors are returned, or, if it is a gradient
palette, only three.

Yes, I wrote this function to save myself the slightest amount of typing.
}
