% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{tmodSummary-class}
\alias{tmodSummary-class}
\title{S4 class for tmodSummary}
\description{
S4 class for tmodSummary
}
\details{
An object of class tmodSummary contains results from multiple
enrichment tests for the same set of gene sets, for example from running
tmodCERNOtest on different contrasts.
}
\seealso{
tmodSummary-data
}
