% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{upset}
\alias{upset}
\title{Upset plot}
\usage{
upset(
  modules,
  mset = NULL,
  min.size = 2,
  min.overlap = 3,
  max.comb = 4,
  min.group = 1,
  value = "number",
  cutoff = NULL,
  labels = NULL
)
}
\arguments{
\item{modules}{optional list of modules for which to make the test}

\item{mset}{Which module set to use. Either a character vector ("LI", "DC" or "all", default: LI) or an object of class tmod (see "Custom module definitions" below)}

\item{min.size}{minimal number of modules in a comparison to show}

\item{min.overlap}{smallest overlap to plot}

\item{max.comb}{Maximum number of combinations}

\item{value}{what to show on the plot: "number" (number of common
elements) or "jaccard" (jaccard index)}
}
\description{
Upset plot
}
