% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Egambia.R
\name{vaccination}
\alias{vaccination}
\title{Transcriptomic responses to vaccination}
\format{
Data frame with one row per gene containing log fold changes and FDR (q
values) for the Fluad vaccine as compared to placebo on day 0, day 1, day 2
and day 3 after the vaccination.
}
\description{
Transcriptomic responses to vaccination
}
\details{
The data shows the time course of transcriptomic responses to influenza
vaccination in healthy volunteers. The source of the data is 
GEO project PRJNA515032, associated with the following paper:

Weiner, January, et al. "Characterization of potential
biomarkers of reactogenicity of licensed antiviral vaccines: randomized
controlled clinical trials conducted by the BIOVACSAFE consortium."
Scientific reports 9.1 (2019): 1-14.

For the data set, 3000 genes with top variance were chosen.
}
