% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/classes2.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,tmod-method}
\alias{[,tmod-method}
\alias{[,tmodGS-method}
\title{Shows the tmod object}
\usage{
\S4method{show}{tmod}(object)

\S4method{[}{tmod}(x, i)

\S4method{show}{tmodGS}(object)

\S4method{[}{tmodGS}(x, i)
}
\arguments{
\item{object}{a tmod object}

\item{x}{a tmod object}

\item{i}{indices specifying elements to extract or replace}
}
\description{
Shows the tmod object

Extracts parts of a tmod object

Shows the tmod object

Extracts parts of a tmod object
}
