% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{getGenes}
\alias{getGenes}
\title{Get genes belonging to a module}
\usage{
getGenes(modules, genelist = NULL, fg = NULL, mset = "LI",
  as.list = FALSE)
}
\arguments{
\item{modules}{module IDs}

\item{genelist}{list of genes}

\item{fg}{genes which are in the foreground set}

\item{mset}{module set to use}

\item{as.list}{should a list of genes rather than a data frame be returned}
}
\value{
data frame containing module to gene mapping, or a list (if
as.list == TRUE
}
\description{
Get genes belonging to a module
}
\details{
Create a data frame mapping each module to a comma separated list of
genes. If genelist is provided, then only genes in that list will be
shown. An optional column, "fg" informs which genes are in the "foreground"
data set.
}
