% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getModuleMembers}
\alias{getModuleMembers}
\title{Return the contents of a module}
\usage{
getModuleMembers(x, mset = "all")
}
\arguments{
\item{x}{a character vector of module names}

\item{mset}{optional, a module set}
}
\value{
A list of modules
}
\description{
Return the contents of a module
}
\details{
This function returns the selected modules from a module set.
}
\examples{
# show the interferon related modules
getModuleMembers(c("LI.M127", "LI.M158.0", "LI.M158.0"))
}
