% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{length}
\alias{length}
\alias{length,tmod-method}
\title{Shows the length of a tmod object}
\usage{
\S4method{length}{tmod}(x)
}
\arguments{
\item{x}{a tmod object}
}
\description{
Shows the length of a tmod object
}
