% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modstats.R
\name{modOverlaps}
\alias{modOverlaps}
\title{Calculate overlaps of the modules}
\usage{
modOverlaps(modules, mset = NULL, stat = "jaccard")
}
\arguments{
\item{modules}{either a character vector with module IDs from mset, or a list which
contains the module members}

\item{mset}{Which module set to use. Either a character vector ("LI", "DC" or "all", default: LI) or an object of class tmod (see "Custom module definitions" below)}

\item{stat}{Type of statistics to return. "jaccard": Jaccard index;
number: number of overlapping genes;
       otherwise, total number of common genes will be returned}
}
\description{
Calculate overlaps of the modules
}
