% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{showModule}
\alias{showModule}
\title{Select genes belonging to a module from a data frame}
\usage{
showModule(x, genes, module, mset = "all", extra = TRUE)
}
\arguments{
\item{x}{a data frame or a vector}

\item{genes}{a character vector with gene IDs}

\item{module}{a single character value, ID of the module to be shown}

\item{mset}{Module set to use; see "tmodUtest" for details}

\item{extra}{If TRUE, additional information about the features will be shown}
}
\value{
Either a filtered vector or (if extra==TRUE) a data frame.
}
\description{
Select genes belonging to a module from a data frame or vector
}
\details{
showModule filters a data frame or a vector such that only genes from a module are
shown. Use it, for example, to show a subset of topTable result from
limma in order to see which genes from a module are significantly
regulated. In essence, this is just a wrapper around "subset()".
}
