% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lem.R
\name{tmodLEA}
\alias{tmodLEA}
\title{Leading Edge Analysis}
\usage{
tmodLEA(l, modules, mset = "all", nodups = TRUE, filter = FALSE)
}
\arguments{
\item{l}{list of genes}

\item{modules}{character vector of module IDs for which to run the LEA}

\item{mset}{Which module set to use. Either a character vector ("LI", "DC" or "all", default: LI) or an object of class tmod}

\item{nodups}{Remove duplicate gene names in l and corresponding
rows from ranks}

\item{filter}{Remove gene names which have no module assignments}
}
\description{
Leading Edge Analysis
}
