% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{upset}
\alias{upset}
\title{Upset plot}
\usage{
upset(
  modules,
  mset = NULL,
  min.size = 2,
  min.overlap = 3,
  max.comb = 4,
  min.group = 1,
  value = "number",
  cutoff = NULL,
  labels = NULL,
  group = TRUE,
  pal = brewer.pal(8, "Dark2")
)
}
\arguments{
\item{modules}{optional list of modules for which to make the test}

\item{mset}{Which module set to use. Either a character vector ("LI", "DC" or "all", default: LI) or an object of class tmod (see "Custom module definitions" below)}

\item{min.size}{minimal number of modules in a comparison to show}

\item{min.overlap}{smallest overlap to plot}

\item{max.comb}{Maximum number of combinations to show (i.e., number of
dots on every vertical segment in the upset plot)}

\item{min.group}{Minimum number of modules in a group. Group with a
smaller number of members will be ignored.}

\item{value}{what to show on the plot: "number" (number of common
elements), "soerensen" (Sørensen–Dice coefficient), 
"overlap" (Szymkiewicz–Simpson coefficient) or "jaccard" (Jaccard index)}

\item{cutoff}{Combinations with the `value` below cutoff will not be shown.}

\item{labels}{Labels for the modules. Character vector with the same
length as `modules`}

\item{group}{Should the modules be grouped by the overlap?}

\item{pal}{Color palette to show the groups.}
}
\description{
Upset plots help to interpret the results of gene set enrichment.
}
