% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Egambia.R
\name{vaccination}
\alias{vaccination}
\title{Transcriptomic responses to vaccination}
\description{
Transcriptomic responses to vaccination
}
\details{
The data shows the time course of transcriptomic responses to influenza
vaccination in healthy volunteers. The source of the data is 
GEO project PRJNA515032, associated with the following paper:

Weiner, January, et al. "Characterization of potential
biomarkers of reactogenicity of licensed antiviral vaccines: randomized
controlled clinical trials conducted by the BIOVACSAFE consortium."
Scientific reports 9.1 (2019): 1-14.
}
